# Reduce merkle tree serialization allocation count