--- a/kvbc/include/merkle_tree_serialization.h
+++ b/kvbc/include/merkle_tree_serialization.h
@@ -24,11 +24,11 @@
 #include "storage/db_types.h"
 #include "string.hpp"
 
+#include <algorithm>
 #include <cstdint>
 #include <iterator>
 #include <limits>
 #include <string>
-
 #include <utility>
 #include <variant>
 #include <vector>
@@ -59,6 +59,11 @@ using BatchedInternalMaskType = std::uint32_t;
 // Serialize integral types in big-endian (network) byte order so that lexicographical comparison works and we can get
 // away without a custom key comparator. toBigEndianStringBuffer() does not allow bools.
 template <typename T>
+void serializeImp(T v, std::string &out) {
+  out += concordUtils::toBigEndianStringBuffer(v);
+}
+
+template <typename T>
 std::string serializeImp(T v) {
   return concordUtils::toBigEndianStringBuffer(v);
 }
@@ -77,59 +82,112 @@ inline std::string serializeImp(storage::v2MerkleTree::detail::EBFTSubtype type)
                      concord::util::toChar(type)};
 }
 
-inline std::string serializeImp(const std::vector<std::uint8_t> &v) {
-  return std::string{std::cbegin(v), std::cend(v)};
+inline void serializeImp(const sparse_merkle::NibblePath &path, std::string &out) {
+  out.append(1, static_cast<std::uint8_t>(path.length()));
+  out.append(std::cbegin(path.data()), std::cend(path.data()));
 }
 
 inline std::string serializeImp(const sparse_merkle::NibblePath &path) {
-  return serializeImp(static_cast<std::uint8_t>(path.length())) + serializeImp(path.data());
+  auto out = std::string{};
+  serializeImp(path, out);
+  return out;
+}
+
+inline void serializeImp(const sparse_merkle::Hash &hash, std::string &out) {
+  out.append(reinterpret_cast<const char *>(hash.data()), hash.size());
 }
 
 inline std::string serializeImp(const sparse_merkle::Hash &hash) {
-  return std::string{reinterpret_cast<const char *>(hash.data()), hash.size()};
+  auto out = std::string{};
+  serializeImp(hash, out);
+  return out;
+}
+
+inline void serializeImp(const sparse_merkle::InternalNodeKey &key, std::string &out) {
+  serializeImp(key.version().value(), out);
+  serializeImp(key.path(), out);
 }
 
 inline std::string serializeImp(const sparse_merkle::InternalNodeKey &key) {
-  return serializeImp(key.version().value()) + serializeImp(key.path());
+  auto out = std::string{};
+  serializeImp(key, out);
+  return out;
+}
+
+inline void serializeImp(const sparse_merkle::LeafKey &key, std::string &out) {
+  serializeImp(key.hash(), out);
+  serializeImp(key.version().value(), out);
 }
 
 inline std::string serializeImp(const sparse_merkle::LeafKey &key) {
-  return serializeImp(key.hash()) + serializeImp(key.version().value());
+  auto out = std::string{};
+  serializeImp(key, out);
+  return out;
+}
+
+inline void serializeImp(const sparse_merkle::LeafChild &child, std::string &out) {
+  serializeImp(child.hash, out);
+  serializeImp(child.key, out);
 }
 
 inline std::string serializeImp(const sparse_merkle::LeafChild &child) {
-  return serializeImp(child.hash) + serializeImp(child.key);
+  auto out = std::string{};
+  serializeImp(child, out);
+  return out;
+}
+
+inline void serializeImp(const sparse_merkle::InternalChild &child, std::string &out) {
+  serializeImp(child.hash, out);
+  serializeImp(child.version.value(), out);
 }
 
 inline std::string serializeImp(const sparse_merkle::InternalChild &child) {
-  return serializeImp(child.hash) + serializeImp(child.version.value());
+  auto out = std::string{};
+  serializeImp(child, out);
+  return out;
 }
 
 inline std::string serializeImp(const sparse_merkle::BatchedInternalNode &intNode) {
   static_assert(sparse_merkle::BatchedInternalNode::MAX_CHILDREN < 32);
 
   struct Visitor {
-    Visitor(std::string &buf) : buf_{buf} {}
+    Visitor(std::string &out) : out_{out} {}
 
     void operator()(const sparse_merkle::LeafChild &leaf) {
-      buf_ += concord::util::toChar(BatchedInternalNodeChildType::Leaf) + serializeImp(leaf);
+      out_ += concord::util::toChar(BatchedInternalNodeChildType::Leaf);
+      serializeImp(leaf, out_);
     }
 
     void operator()(const sparse_merkle::InternalChild &internal) {
-      buf_ += concord::util::toChar(BatchedInternalNodeChildType::Internal) + serializeImp(internal);
+      out_ += concord::util::toChar(BatchedInternalNodeChildType::Internal);
+      serializeImp(internal, out_);
     }
 
-    std::string &buf_;
+    std::string &out_;
   };
 
-  const auto &children = intNode.children();
-  std::string serializedChildren;
   auto mask = BatchedInternalMaskType{0};
+
+  // We assume the largest batched internal node contains with 15 internal and 16 leaf children. We then factor in the
+  // mask that indicates which child is set and the single-byte child type bytes serialized for every child.
+  auto out = std::string{};
+  out.reserve((sparse_merkle::InternalChild::SIZE_IN_BYTES * 15) + (sparse_merkle::LeafChild::SIZE_IN_BYTES * 16) +
+              sizeof(mask) + sparse_merkle::BatchedInternalNode::MAX_CHILDREN);
+
+  const auto &children = intNode.children();
+
+  // Serialize by first inserting a bitmask of type BatchedInternalMaskType specifying which indexes are set in the
+  // children array. Then, serialize a single byte type before each child and then the child itself - done by the
+  // visitor.
+  //
+  // Start by serializing an empty (all zeroes) mask. Afterwards, overwrite it when the actual mask is calculated. This
+  // allows us to loop once only and avoid shifting the data in the output buffer.
+  std::fill_n(std::back_inserter(out), sizeof(mask), 0);
   for (auto i = 0u; i < children.size(); ++i) {
     const auto &child = children[i];
     if (child) {
       mask |= (1 << i);
-      std::visit(Visitor{serializedChildren}, *child);
+      std::visit(Visitor{out}, *child);
     }
   }
 
@@ -138,9 +196,16 @@ inline std::string serializeImp(const sparse_merkle::BatchedInternalNode &intNod
     return std::string{};
   }
 
-  // Serialize by putting a bitmask (of type BatchedInternalMaskType) specifying which indexes are set in the children
-  // array. Then, put a type before each child and then the child itself.
-  return serializeImp(mask) + serializedChildren;
+  // Set the actual mask in the output.
+  {
+    const auto serializedMask = serializeImp(mask);
+    ConcordAssertEQ(sizeof(mask), serializedMask.size());
+    auto i = 0;
+    for (auto c : serializedMask) {
+      out[i++] = c;
+    }
+  }
+  return out;
 }
 
 inline std::string serializeImp(const block::detail::Node &node) {