Use an output parameter instead of std::string concatenation when
serializing merkle tree internals. That has the effect of reducing the
count of memory allocations. Special attention is given to serialization
of BatchedInternalNodes as it was creating a big number of temporary
strings that were large enough to trigger allocations (without
taking advantage of small string optimizations).

Add a benchmark for BatchedInternalNodes. Change constants for block
count and key/value count and sizes in order to better reflect real uses
cases.

```
Before:
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
serializeBatchedInternalNode         3229 ns         3229 ns       212506
```

```
After:
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
serializeBatchedInternalNode          680 ns          680 ns       818500
```

Thanks to @andrewjstone for observing and measuring.